+++ lib/modules/manager/api.ts
@@ -89,7 +89,7 @@ import * as pub from './pub/index.ts';
 import * as puppet from './puppet/index.ts';
 import * as pyenv from './pyenv/index.ts';
 import * as quadlet from './quadlet/index.ts';
-import * as renovateConfigPresets from './renovate-config-presets/index.ts';
+import * as renovateConfig from './renovate-config/index.ts';
 import * as rubyVersion from './ruby-version/index.ts';
 import * as runtimeVersion from './runtime-version/index.ts';
 import * as sbt from './sbt/index.ts';
@@ -206,7 +206,7 @@ api.set('pub', pub);
 api.set('puppet', puppet);
 api.set('pyenv', pyenv);
 api.set('quadlet', quadlet);
-api.set('renovate-config-presets', renovateConfigPresets);
+api.set('renovate-config', renovateConfig);
 api.set('ruby-version', rubyVersion);
 api.set('runtime-version', runtimeVersion);
 api.set('sbt', sbt);
