+++ lib/modules/manager/renovate-config/index.ts
