+++ lib/modules/manager/renovate-config/schema.ts
